Add sphering to the main workflow
